windows: use the known folders API
On Windows, the actual path for the local and roaming AppData folders is
not guaranteed to be the same as its environment variable equivalent.
The environment variables are just default values.

The actual paths should be get with the Known Folders Windows API.

Fortunately the x/sys Go repository abstracts all the Win32 API calls so
that using the correct API is a small and simple change.

This won't change the output on Windows systems with standard paths.

Closes: #1